test: move tests for `Vault` class to a separate file
This commit moves the tests to add and find vaults to a separate
file. It also updates the tests to invoke `::add` and
`::find_by_vault_name` to add and find vaults from the `vaults` table.

Getter methods are also defined in the `Vault` class so that the
instance variables can be referenced for their values to be verified.
