+++ test/password_manager_test.rb
@@ -152,35 +152,4 @@ class PasswordManagerTest < Minitest::Test
     assert vault2
     assert_equal "1", vault2["id"]
   end
-
-  def test_add_vault_with_valid_user_id
-    @storage.add_user("admin", "123")
-    @storage.add_vault(1, "My Vault")
-
-    vault = @storage.find_vault(1, "My Vault")
-    assert vault
-    assert_equal "1", vault["user_id"]
-    assert_equal "My Vault", vault["name"]
-  end
-
-  def test_add_vault_without_valid_user_id
-    assert_raises PG::ForeignKeyViolation do
-      @storage.add_vault(1, "My Vault")
-    end
-
-    assert_nil @storage.find_vault(1, "My Vault")
-  end
-
-  def test_find_vault_with_valid_user_id
-    @storage.add_user("admin", "123")
-    @storage.add_vault(1, "My Vault")
-
-    assert @storage.find_vault(1, "My Vault")
-    assert @storage.find_vault(1, "my vault")
-    assert_nil @storage.find_vault(1, "test vault")
-  end
-
-  def test_find_vault_without_valid_user_id
-    assert_nil @storage.find_vault(1, "my vault")
-  end
 end
